+++ hydra-cluster/test/Test/EndToEndSpec.hs
@@ -678,6 +678,7 @@ initAndClose tmpDir tracer clusterIx hydraScriptsTxId node@RunningNode{nodeSocke
                   , "datum" .= Null
                   , "datumhash" .= Null
                   , "inlineDatum" .= Null
+                  , "inlineDatumRaw" .= Null
                   , "referenceScript" .= Null
                   ]
               )
@@ -689,6 +690,7 @@ initAndClose tmpDir tracer clusterIx hydraScriptsTxId node@RunningNode{nodeSocke
                   , "datum" .= Null
                   , "datumhash" .= Null
                   , "inlineDatum" .= Null
+                  , "inlineDatumRaw" .= Null
                   , "referenceScript" .= Null
                   ]
               )